Version 1.1
Before; Sudoku code looked for where individual cell's checkers only had one possible solution based on the numbers that see it from its row and column, and block. 

Now; Sudoku code looks within rows, columns and blocks where a clue for the cell is mentioned only once, i.e.: if the number 'N' only appears in the cell's checkers once in that whole row, even if that cell's checkers present many possibilities, that 'N' is filled as the true answer.